requirements.txt
openai==2.9.0
psycopg2-binary==2.9.11
dotenv==0.9.9
pgvector-0.4.2